requirements.txt
pymupdf>=1.24.0
python-docx>=1.1.0
sentence-transformers>=2.6.0
torch>=2.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
rank-bm25>=0.2.2
openai>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pytest>=8.0.0

# Install this repo as a package so `import app` works from any directory (fixes ModuleNotFoundError).
# From the project root: pip install -e .
-e .
